Extract WTA prompt building and turn metrics
Move ACP prompt assembly, pane context resolution, and prompt audit logging into focused prompt modules. Move per-turn timing and response telemetry into a neutral metrics module while preserving session and callback orchestration in the ACP client.


Copilot-Session: f0b2a188-f62a-48b7-b316-6fc31ed3880c
