+++ tools/wta/src/protocol/acp/mod.rs
@@ -4,7 +4,9 @@ pub mod failure;
 pub mod model_select;
 pub mod probe;
 pub mod prompt;
+pub(crate) mod prompt_builder;
 pub(crate) mod prompt_context;
 pub(crate) mod session_list;
 pub mod soft_stop;
 pub(crate) mod spawn;
+pub(crate) mod turn_metrics;
